aws: Move the agent-side ENI IPAM logic to `pkg/aws/agent`
This commit moves all the AWS-specific logic that was previously in the
shared IPAM package (in `pkg/ipam/eni.go`) to a new `pkg/aws/agent`
package. The underlying logic is unchanged, only the code moves, and it
now gets wired via hive as an `ipam.CloudProvider`.

Also move `pkg/ipam.parsePrimaryMAC` to `pkg/mac.ParseMACOrUnset`. This
allows this function to be used from the new `pkg/aws/agent` package
without having it import `pkg/ipam`, also `pkg/mac` is a more fitting
home for it.
