--- pkg/ipam/eni_test.go
@@ -1,872 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright Authors of Cilium
-
-package ipam
-
-import (
-	"context"
-	"log/slog"
-	"net/netip"
-	"testing"
-	"time"
-
-	"github.com/cilium/hive/cell"
-	"github.com/cilium/hive/hivetest"
-	"github.com/cilium/hive/job"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	"github.com/cilium/cilium/daemon/k8s"
-	awsTypes "github.com/cilium/cilium/pkg/aws/types"
-	"github.com/cilium/cilium/pkg/defaults"
-	"github.com/cilium/cilium/pkg/hive"
-	iputil "github.com/cilium/cilium/pkg/ip"
-	ipamTypes "github.com/cilium/cilium/pkg/ipam/types"
-	"github.com/cilium/cilium/pkg/ipmasq"
-	ciliumv2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
-	k8sClient "github.com/cilium/cilium/pkg/k8s/client/testutils"
-	"github.com/cilium/cilium/pkg/mac"
-	"github.com/cilium/cilium/pkg/node"
-	nodeTypes "github.com/cilium/cilium/pkg/node/types"
-	"github.com/cilium/cilium/pkg/option"
-	"github.com/cilium/cilium/pkg/slices"
-)
-
-// addrs converts string IPs to []iputil.Addr for use in tests.
-func addrs(ips ...string) []iputil.Addr {
-	return slices.Map(ips, func(s string) iputil.Addr {
-		return iputil.AddrFrom(netip.MustParseAddr(s))
-	})
-}
-
-// prefixes converts string CIDRs to []iputil.Prefix for use in tests.
-func prefixes(ps ...string) []iputil.Prefix {
-	return slices.Map(ps, func(s string) iputil.Prefix {
-		return iputil.PrefixFrom(netip.MustParsePrefix(s))
-	})
-}
-
-func Test_validateENIConfig(t *testing.T) {
-	type args struct {
-		node *ciliumv2.CiliumNode
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-		want    string
-	}{
-		{
-			name: "Consistent ENI config",
-			args: args{
-				node: &ciliumv2.CiliumNode{
-					Spec: ciliumv2.NodeSpec{
-						IPAM: ipamTypes.IPAMSpec{
-							Pool: ipamTypes.AllocationMap{
-								"10.1.1.226": ipamTypes.AllocationIP{
-									Resource: "eni-1",
-								},
-							},
-						},
-					},
-					Status: ciliumv2.NodeStatus{
-						ENI: awsTypes.ENIStatus{
-							ENIs: map[string]awsTypes.ENI{
-								"eni-1": {
-									ID: "eni-1",
-									IP: iputil.AddrFrom(netip.MustParseAddr("10.1.1.225")),
-									Subnet: awsTypes.AwsSubnet{
-										CIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.1.0/24")),
-									},
-									Addresses: addrs(
-										"10.1.1.226",
-										"10.1.1.229",
-									),
-									VPC: awsTypes.AwsVPC{
-										ID:          "vpc-1",
-										PrimaryCIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.0.0/16")),
-										CIDRs: prefixes(
-											"10.1.0.0/16",
-											"10.2.0.0/16",
-										),
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "Missing VPC Primary CIDR",
-			args: args{
-				node: &ciliumv2.CiliumNode{
-					Spec: ciliumv2.NodeSpec{
-						IPAM: ipamTypes.IPAMSpec{
-							Pool: ipamTypes.AllocationMap{
-								"10.1.1.226": ipamTypes.AllocationIP{
-									Resource: "eni-1",
-								},
-							},
-						},
-					},
-					Status: ciliumv2.NodeStatus{
-						ENI: awsTypes.ENIStatus{
-							ENIs: map[string]awsTypes.ENI{
-								"eni-1": {
-									ID: "eni-1",
-									IP: iputil.AddrFrom(netip.MustParseAddr("10.1.1.225")),
-									Subnet: awsTypes.AwsSubnet{
-										CIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.1.0/24")),
-									},
-									Addresses: addrs(
-										"10.1.1.226",
-										"10.1.1.229",
-									),
-									VPC: awsTypes.AwsVPC{
-										ID: "vpc-1",
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			wantErr: true,
-			want:    "VPC Primary CIDR not set for ENI eni-1",
-		},
-		{
-			name: "VPC CIDRs contain invalid value",
-			args: args{
-				node: &ciliumv2.CiliumNode{
-					Spec: ciliumv2.NodeSpec{
-						IPAM: ipamTypes.IPAMSpec{
-							Pool: ipamTypes.AllocationMap{
-								"10.1.1.226": ipamTypes.AllocationIP{
-									Resource: "eni-1",
-								},
-							},
-						},
-					},
-					Status: ciliumv2.NodeStatus{
-						ENI: awsTypes.ENIStatus{
-							ENIs: map[string]awsTypes.ENI{
-								"eni-1": {
-									ID: "eni-1",
-									IP: iputil.AddrFrom(netip.MustParseAddr("10.1.1.225")),
-									Subnet: awsTypes.AwsSubnet{
-										CIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.1.0/24")),
-									},
-									Addresses: addrs(
-										"10.1.1.226",
-										"10.1.1.229",
-									),
-									VPC: awsTypes.AwsVPC{
-										ID:          "vpc-1",
-										PrimaryCIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.0.0/16")),
-										CIDRs: []iputil.Prefix{
-											{},
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			wantErr: true,
-			want:    "VPC CIDR not set for ENI eni-1",
-		},
-		{
-			name: "ENI not found in status",
-			args: args{
-				node: &ciliumv2.CiliumNode{
-					Spec: ciliumv2.NodeSpec{
-						IPAM: ipamTypes.IPAMSpec{
-							Pool: ipamTypes.AllocationMap{
-								"10.1.1.226": ipamTypes.AllocationIP{
-									Resource: "eni-1",
-								},
-							},
-						},
-					},
-					Status: ciliumv2.NodeStatus{
-						ENI: awsTypes.ENIStatus{
-							ENIs: map[string]awsTypes.ENI{
-								"eni-2": {
-									ID: "eni-2",
-									IP: iputil.AddrFrom(netip.MustParseAddr("10.1.1.225")),
-									Subnet: awsTypes.AwsSubnet{
-										CIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.1.0/24")),
-									},
-									Addresses: addrs(
-										"10.1.1.226",
-										"10.1.1.229",
-									),
-									VPC: awsTypes.AwsVPC{
-										ID:          "vpc-1",
-										PrimaryCIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.0.0/16")),
-										CIDRs: prefixes(
-											"10.1.0.0/16",
-											"10.2.0.0/16",
-										),
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			wantErr: true,
-			want:    "ENI eni-1 not found in status",
-		},
-		{
-			name: "ENI IP not found in status",
-			args: args{
-				node: &ciliumv2.CiliumNode{
-					Spec: ciliumv2.NodeSpec{
-						IPAM: ipamTypes.IPAMSpec{
-							Pool: ipamTypes.AllocationMap{
-								"10.1.1.227": ipamTypes.AllocationIP{
-									Resource: "eni-1",
-								},
-							},
-						},
-					},
-					Status: ciliumv2.NodeStatus{
-						ENI: awsTypes.ENIStatus{
-							ENIs: map[string]awsTypes.ENI{
-								"eni-1": {
-									ID: "eni-1",
-									IP: iputil.AddrFrom(netip.MustParseAddr("10.1.1.225")),
-									Subnet: awsTypes.AwsSubnet{
-										CIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.1.0/24")),
-									},
-									Addresses: addrs(
-										"10.1.1.226",
-										"10.1.1.229",
-									),
-									VPC: awsTypes.AwsVPC{
-										ID:          "vpc-1",
-										PrimaryCIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.0.0/16")),
-										CIDRs: prefixes(
-											"10.1.0.0/16",
-											"10.2.0.0/16",
-										),
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			wantErr: true,
-			want:    "ENI eni-1 does not have address 10.1.1.227",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := validateENIConfig(tt.args.node)
-			require.Equal(t, tt.wantErr, got != nil, "error: %v", got)
-			if tt.wantErr {
-				require.Equal(t, tt.want, got.Error())
-			}
-		})
-	}
-}
-
-func TestBuildENIAllocationResult(t *testing.T) {
-	node := &ciliumv2.CiliumNode{}
-	node.Status.ENI.ENIs = map[string]awsTypes.ENI{
-		"eni-1": {
-			ID:  "eni-1",
-			MAC: "aa:bb:cc:dd:ee:01",
-			Addresses: addrs(
-				"10.1.1.10",
-				"10.1.1.11",
-			),
-			Number: 1,
-			Subnet: awsTypes.AwsSubnet{
-				CIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.1.0/24")),
-			},
-			VPC: awsTypes.AwsVPC{
-				PrimaryCIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.0.0/16")),
-				CIDRs:       prefixes("10.2.0.0/16"),
-			},
-		},
-		"eni-2": {
-			ID:  "eni-2",
-			MAC: "aa:bb:cc:dd:ee:02",
-			Addresses: addrs(
-				"10.3.1.20",
-			),
-			Number: 2,
-			Subnet: awsTypes.AwsSubnet{
-				CIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.3.1.0/24")),
-			},
-			VPC: awsTypes.AwsVPC{
-				PrimaryCIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.0.0/16")),
-				CIDRs:       prefixes("10.2.0.0/16"),
-			},
-		},
-	}
-
-	conf := &option.DaemonConfig{}
-	logger := hivetest.Logger(t)
-
-	t.Run("secondary IP on eni-1", func(t *testing.T) {
-		result, err := buildENIAllocationResult(logger, netip.MustParseAddr("10.1.1.10"), "", node.Status.ENI.ENIs, conf, nil)
-		require.NoError(t, err)
-		require.Equal(t, mac.MustParseMAC("aa:bb:cc:dd:ee:01"), result.PrimaryMAC)
-		require.Equal(t, "1", result.InterfaceNumber)
-		require.Equal(t, netip.MustParseAddr("10.1.1.1"), result.GatewayIP)
-		require.Contains(t, result.CIDRs, netip.MustParsePrefix("10.1.0.0/16"))
-		require.Contains(t, result.CIDRs, netip.MustParsePrefix("10.2.0.0/16"))
-	})
-
-	t.Run("secondary IP on eni-2", func(t *testing.T) {
-		result, err := buildENIAllocationResult(logger, netip.MustParseAddr("10.3.1.20"), "", node.Status.ENI.ENIs, conf, nil)
-		require.NoError(t, err)
-		require.Equal(t, mac.MustParseMAC("aa:bb:cc:dd:ee:02"), result.PrimaryMAC)
-		require.Equal(t, "2", result.InterfaceNumber)
-		require.Equal(t, netip.MustParseAddr("10.3.1.1"), result.GatewayIP)
-	})
-
-	t.Run("unknown IP returns error", func(t *testing.T) {
-		_, err := buildENIAllocationResult(logger, netip.MustParseAddr("10.99.99.99"), "", node.Status.ENI.ENIs, conf, nil)
-		require.Error(t, err)
-		require.Contains(t, err.Error(), "unable to find ENI for IP")
-	})
-
-	t.Run("native routing CIDR is appended", func(t *testing.T) {
-		confWithNative := &option.DaemonConfig{
-			EnableIPv4:            true,
-			IPv4NativeRoutingCIDR: netip.MustParsePrefix("10.0.0.0/8"),
-		}
-		result, err := buildENIAllocationResult(logger, netip.MustParseAddr("10.1.1.10"), "", node.Status.ENI.ENIs, confWithNative, nil)
-		require.NoError(t, err)
-		require.Contains(t, result.CIDRs, netip.MustParsePrefix("10.0.0.0/8"))
-	})
-
-	t.Run("IPv4 native routing CIDR is not appended when IPv4 is disabled", func(t *testing.T) {
-		confWithNative := &option.DaemonConfig{
-			IPv4NativeRoutingCIDR: netip.MustParsePrefix("10.0.0.0/8"),
-		}
-		result, err := buildENIAllocationResult(logger, netip.MustParseAddr("10.1.1.10"), "", node.Status.ENI.ENIs, confWithNative, nil)
-		require.NoError(t, err)
-		require.NotContains(t, result.CIDRs, netip.MustParsePrefix("10.0.0.0/8"))
-	})
-}
-
-func TestBuildENIAllocationResultPrefixDelegation(t *testing.T) {
-	node := &ciliumv2.CiliumNode{}
-	node.Status.ENI.ENIs = map[string]awsTypes.ENI{
-		"eni-1": {
-			ID:  "eni-1",
-			MAC: "aa:bb:cc:dd:ee:01",
-			Prefixes: prefixes(
-				"10.1.1.0/28",
-				"10.1.1.16/28",
-			),
-			IPv6Prefixes: prefixes(
-				"2001:db8::/80",
-			),
-			Number: 1,
-			Subnet: awsTypes.AwsSubnet{
-				CIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.1.0/24")),
-			},
-			VPC: awsTypes.AwsVPC{
-				PrimaryCIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.0.0/16")),
-			},
-		},
-	}
-
-	conf := &option.DaemonConfig{}
-	logger := hivetest.Logger(t)
-
-	t.Run("IP in first prefix", func(t *testing.T) {
-		result, err := buildENIAllocationResult(logger, netip.MustParseAddr("10.1.1.5"), "", node.Status.ENI.ENIs, conf, nil)
-		require.NoError(t, err)
-		require.Equal(t, mac.MustParseMAC("aa:bb:cc:dd:ee:01"), result.PrimaryMAC)
-		require.Equal(t, "1", result.InterfaceNumber)
-	})
-
-	t.Run("IP in second prefix", func(t *testing.T) {
-		result, err := buildENIAllocationResult(logger, netip.MustParseAddr("10.1.1.20"), "", node.Status.ENI.ENIs, conf, nil)
-		require.NoError(t, err)
-		require.Equal(t, mac.MustParseMAC("aa:bb:cc:dd:ee:01"), result.PrimaryMAC)
-	})
-
-	t.Run("IP outside all prefixes", func(t *testing.T) {
-		_, err := buildENIAllocationResult(logger, netip.MustParseAddr("10.1.1.32"), "", node.Status.ENI.ENIs, conf, nil)
-		require.Error(t, err)
-	})
-
-	t.Run("IP in IPv6 prefix", func(t *testing.T) {
-		result, err := buildENIAllocationResult(logger, netip.MustParseAddr("2001:db8::1"), "", node.Status.ENI.ENIs, conf, nil)
-		require.NoError(t, err)
-		require.Equal(t, mac.MustParseMAC("aa:bb:cc:dd:ee:01"), result.PrimaryMAC)
-		require.Equal(t, "1", result.InterfaceNumber)
-		require.Equal(t, netip.MustParseAddr("fe80:ec2::1"), result.GatewayIP)
-	})
-
-	t.Run("IPv6 native routing CIDR is appended", func(t *testing.T) {
-		confWithNative := &option.DaemonConfig{
-			EnableIPv6:            true,
-			IPv6NativeRoutingCIDR: netip.MustParsePrefix("2001:db8::/64"),
-		}
-		result, err := buildENIAllocationResult(logger, netip.MustParseAddr("2001:db8::1"), "", node.Status.ENI.ENIs, confWithNative, nil)
-		require.NoError(t, err)
-		require.Contains(t, result.CIDRs, netip.MustParsePrefix("2001:db8::/64"))
-	})
-
-	t.Run("IPv6 native routing CIDR is not appended when IPv6 is disabled", func(t *testing.T) {
-		confWithNative := &option.DaemonConfig{
-			IPv6NativeRoutingCIDR: netip.MustParsePrefix("2001:db8::/64"),
-		}
-		result, err := buildENIAllocationResult(logger, netip.MustParseAddr("2001:db8::1"), "", node.Status.ENI.ENIs, confWithNative, nil)
-		require.NoError(t, err)
-		require.NotContains(t, result.CIDRs, netip.MustParsePrefix("2001:db8::/64"))
-	})
-}
-
-func TestBuildENIAllocationResultIPMasq(t *testing.T) {
-	enis := map[string]awsTypes.ENI{
-		"eni-1": {
-			ID: "eni-1",
-			Addresses: addrs(
-				"10.1.1.226",
-				"10.1.1.229",
-			),
-			Subnet: awsTypes.AwsSubnet{
-				CIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.1.0/24")),
-			},
-			VPC: awsTypes.AwsVPC{
-				ID:          "vpc-1",
-				PrimaryCIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.0.0/16")),
-				CIDRs: prefixes(
-					"10.2.0.0/16",
-				),
-			},
-		},
-	}
-
-	conf := &option.DaemonConfig{EnableIPMasqAgent: true}
-	ipMasqAgent := ipmasq.NewIPMasqAgent(hivetest.Logger(t), "", ipMasqMapDummy{})
-	require.NoError(t, ipMasqAgent.Start())
-	defer ipMasqAgent.Stop()
-
-	result, err := buildENIAllocationResult(hivetest.Logger(t), netip.MustParseAddr("10.1.1.226"), "", enis, conf, ipMasqAgent)
-	require.NoError(t, err)
-	// The resulting CIDRs should contain the VPC CIDRs and the default
-	// ip-masq-agent CIDRs from pkg/ipmasq/ipmasq.go
-	require.ElementsMatch(
-		t,
-		[]netip.Prefix{
-			// VPC CIDRs
-			netip.MustParsePrefix("10.1.0.0/16"),
-			netip.MustParsePrefix("10.2.0.0/16"),
-			// Default ip-masq-agent CIDRs
-			netip.MustParsePrefix("10.0.0.0/8"),
-			netip.MustParsePrefix("172.16.0.0/12"),
-			netip.MustParsePrefix("192.168.0.0/16"),
-			netip.MustParsePrefix("100.64.0.0/10"),
-			netip.MustParsePrefix("192.0.0.0/24"),
-			netip.MustParsePrefix("192.0.2.0/24"),
-			netip.MustParsePrefix("192.88.99.0/24"),
-			netip.MustParsePrefix("198.18.0.0/15"),
-			netip.MustParsePrefix("198.51.100.0/24"),
-			netip.MustParsePrefix("203.0.113.0/24"),
-			netip.MustParsePrefix("240.0.0.0/4"),
-			netip.MustParsePrefix("169.254.0.0/16"),
-		},
-		result.CIDRs,
-	)
-}
-
-func TestBuildENIAllocationResultIPMasqIPv6(t *testing.T) {
-	enis := map[string]awsTypes.ENI{
-		"eni-1": {
-			ID:           "eni-1",
-			IPv6Prefixes: prefixes("2001:db8::/80"),
-			Subnet: awsTypes.AwsSubnet{
-				CIDR: iputil.PrefixFrom(netip.MustParsePrefix("2001:db8::/64")),
-			},
-		},
-	}
-
-	conf := &option.DaemonConfig{EnableIPv6: true, EnableIPMasqAgent: true}
-	ipMasqAgent := ipmasq.NewIPMasqAgent(hivetest.Logger(t), "", ipMasqMapDummy{})
-	require.NoError(t, ipMasqAgent.Start())
-	defer ipMasqAgent.Stop()
-
-	result, err := buildENIAllocationResult(hivetest.Logger(t), netip.MustParseAddr("2001:db8::1"), "", enis, conf, ipMasqAgent)
-	require.NoError(t, err)
-	// Only the IPv6 ip-masq-agent CIDRs should be appended; the family
-	// filter in buildENIAllocationResult must exclude the IPv4 defaults.
-	require.Contains(t, result.CIDRs, netip.MustParsePrefix("fe80::/10"))
-	for _, c := range result.CIDRs {
-		require.True(t, c.Addr().Is6(), "unexpected IPv4 CIDR %s for an IPv6 allocation", c)
-	}
-}
-
-func TestEniContainsIP(t *testing.T) {
-	eni := awsTypes.ENI{
-		IP:           iputil.AddrFrom(netip.MustParseAddr("10.0.0.100")),
-		Addresses:    addrs("10.0.0.1", "10.0.0.2"),
-		Prefixes:     prefixes("10.0.1.0/28"),
-		IPv6Prefixes: prefixes("2001:db8::/80"),
-	}
-
-	// Primary IP match
-	require.True(t, eniContainsIP(eni, netip.MustParseAddr("10.0.0.100")))
-
-	// Secondary address match
-	require.True(t, eniContainsIP(eni, netip.MustParseAddr("10.0.0.1")))
-	require.True(t, eniContainsIP(eni, netip.MustParseAddr("10.0.0.2")))
-	require.False(t, eniContainsIP(eni, netip.MustParseAddr("10.0.0.3")))
-
-	// Prefix match
-	require.True(t, eniContainsIP(eni, netip.MustParseAddr("10.0.1.0")))
-	require.True(t, eniContainsIP(eni, netip.MustParseAddr("10.0.1.15")))
-	require.False(t, eniContainsIP(eni, netip.MustParseAddr("10.0.1.16")))
-
-	// IPv6 prefix match
-	require.True(t, eniContainsIP(eni, netip.MustParseAddr("2001:db8::")))
-	require.True(t, eniContainsIP(eni, netip.MustParseAddr("2001:db8::1")))
-	require.False(t, eniContainsIP(eni, netip.MustParseAddr("2001:db8:0:0:1::")))
-
-	// Empty ENI
-	require.False(t, eniContainsIP(awsTypes.ENI{}, netip.MustParseAddr("10.0.0.1")))
-}
-
-func TestAddressCoveredByPrefix(t *testing.T) {
-	prefixes := []netip.Prefix{
-		netip.MustParsePrefix("10.0.0.0/24"),
-		netip.MustParsePrefix("192.168.1.0/28"),
-	}
-
-	tests := []struct {
-		name     string
-		addr     netip.Addr
-		prefixes []netip.Prefix
-		want     bool
-	}{
-		{"address in first prefix", netip.MustParseAddr("10.0.0.50"), prefixes, true},
-		{"address in second prefix", netip.MustParseAddr("192.168.1.10"), prefixes, true},
-		{"address outside all prefixes", netip.MustParseAddr("172.16.0.1"), prefixes, false},
-		{"empty prefixes", netip.MustParseAddr("10.0.0.1"), nil, false},
-		{"boundary - first address", netip.MustParseAddr("10.0.0.0"), prefixes, true},
-		{"boundary - last address", netip.MustParseAddr("10.0.0.255"), prefixes, true},
-		{"boundary - just outside", netip.MustParseAddr("10.0.1.0"), prefixes, false},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			require.Equal(t, tt.want, addressCoveredByPrefix(tt.addr, tt.prefixes))
-		})
-	}
-}
-
-func TestENIPoolsAccessorFromResource(t *testing.T) {
-	t.Run("no ENIs returns spec pools", func(t *testing.T) {
-		node := &ciliumv2.CiliumNode{}
-		result := eniPoolAccessor.FromResource(node)
-		require.Empty(t, result.Allocated)
-	})
-
-	t.Run("secondary IPs as /32 CIDRs", func(t *testing.T) {
-		node := &ciliumv2.CiliumNode{}
-		node.Status.ENI.ENIs = map[string]awsTypes.ENI{
-			"eni-1": {
-				Addresses: addrs("10.0.0.1", "10.0.0.2"),
-			},
-		}
-
-		result := eniPoolAccessor.FromResource(node)
-		require.Len(t, result.Allocated, 1)
-		require.Equal(t, defaults.IPAMDefaultIPPool, result.Allocated[0].Pool)
-		require.True(t, result.Allocated[0].AllowFirstIP)
-		require.True(t, result.Allocated[0].AllowLastIP)
-		require.Contains(t, result.Allocated[0].CIDRs, iputil.PrefixFrom(netip.MustParsePrefix("10.0.0.1/32")))
-		require.Contains(t, result.Allocated[0].CIDRs, iputil.PrefixFrom(netip.MustParsePrefix("10.0.0.2/32")))
-	})
-
-	t.Run("prefix delegation writes prefixes and excludes covered addresses", func(t *testing.T) {
-		node := &ciliumv2.CiliumNode{}
-		node.Status.ENI.ENIs = map[string]awsTypes.ENI{
-			"eni-1": {
-				// Mimics the pkg/aws/api.parseENI behavior: Addresses contains the ENI secondary
-				// IPs, the ENI primary if UsePrimaryAddress and the 16 IPs expanded from the /28 prefix.
-				// Prefixes contains the raw /28.
-				Addresses: addrs(
-					"10.0.0.1", // ENI primary IP (UsePrimaryAddress)
-					"10.0.0.16", "10.0.0.17", "10.0.0.18", "10.0.0.19",
-					"10.0.0.20", "10.0.0.21", "10.0.0.22", "10.0.0.23",
-					"10.0.0.24", "10.0.0.25", "10.0.0.26", "10.0.0.27",
-					"10.0.0.28", "10.0.0.29", "10.0.0.30", "10.0.0.31",
-				),
-				Prefixes: prefixes("10.0.0.16/28"),
-			},
-		}
-
-		result := eniPoolAccessor.FromResource(node)
-		require.Len(t, result.Allocated, 1)
-		require.Equal(t, defaults.IPAMDefaultIPPool, result.Allocated[0].Pool)
-		require.True(t, result.Allocated[0].AllowFirstIP)
-		require.True(t, result.Allocated[0].AllowLastIP)
-		// Should contain the /28 prefix and the primary IP as /32,
-		// but not the 16 expanded prefix IPs.
-		require.Contains(t, result.Allocated[0].CIDRs, iputil.PrefixFrom(netip.MustParsePrefix("10.0.0.16/28")))
-		require.Contains(t, result.Allocated[0].CIDRs, iputil.PrefixFrom(netip.MustParsePrefix("10.0.0.1/32")))
-		require.Len(t, result.Allocated[0].CIDRs, 2) // 1 prefix + 1 primary IP
-	})
-
-	t.Run("IPv6 prefixes are advertised as pool CIDRs", func(t *testing.T) {
-		node := &ciliumv2.CiliumNode{}
-		node.Status.ENI.ENIs = map[string]awsTypes.ENI{
-			"eni-1": {
-				Addresses:    addrs("10.0.0.1"),
-				Prefixes:     prefixes("10.0.0.16/28"),
-				IPv6Prefixes: prefixes("2001:db8::/80"),
-			},
-		}
-
-		result := eniPoolAccessor.FromResource(node)
-		require.Len(t, result.Allocated, 1)
-		require.Contains(t, result.Allocated[0].CIDRs, iputil.PrefixFrom(netip.MustParsePrefix("10.0.0.16/28")))
-		require.Contains(t, result.Allocated[0].CIDRs, iputil.PrefixFrom(netip.MustParsePrefix("10.0.0.1/32")))
-		require.Contains(t, result.Allocated[0].CIDRs, iputil.PrefixFrom(netip.MustParsePrefix("2001:db8::/80")))
-	})
-
-	t.Run("excluded ENIs are skipped", func(t *testing.T) {
-		node := &ciliumv2.CiliumNode{}
-		node.Spec.ENI.ExcludeInterfaceTags = map[string]string{"skip": "true"}
-		node.Status.ENI.ENIs = map[string]awsTypes.ENI{
-			"eni-1": {
-				Addresses: addrs("10.0.0.1"),
-				Tags:      map[string]string{"skip": "true"},
-			},
-			"eni-2": {
-				Addresses: addrs("10.0.0.2"),
-			},
-		}
-
-		result := eniPoolAccessor.FromResource(node)
-		require.Len(t, result.Allocated, 1)
-		require.True(t, result.Allocated[0].AllowFirstIP)
-		require.True(t, result.Allocated[0].AllowLastIP)
-		require.Len(t, result.Allocated[0].CIDRs, 1)
-		require.Contains(t, result.Allocated[0].CIDRs, iputil.PrefixFrom(netip.MustParsePrefix("10.0.0.2/32")))
-	})
-}
-
-func TestENIMultiPoolAllocator(t *testing.T) {
-	eniCIDR1 := netip.MustParsePrefix("10.0.10.1/32")
-	eniCIDR2 := netip.MustParsePrefix("10.0.10.2/32")
-	initialNode := &ciliumv2.CiliumNode{
-		ObjectMeta: metav1.ObjectMeta{Name: nodeTypes.GetName()},
-		Status: ciliumv2.NodeStatus{
-			ENI: awsTypes.ENIStatus{
-				ENIs: map[string]awsTypes.ENI{
-					"eni-1": {
-						Addresses: addrs(eniCIDR1.Addr().String(), eniCIDR2.Addr().String()),
-						VPC: awsTypes.AwsVPC{
-							PrimaryCIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.0.0.0/16")),
-						},
-					},
-				},
-			},
-		},
-	}
-
-	var (
-		jg             job.Group
-		localNode      k8s.LocalCiliumNodeResource
-		localNodeStore *node.LocalNodeStore
-		clientset      *k8sClient.FakeClientset
-	)
-	h := hive.New(
-		k8s.ResourcesCell,
-		k8sClient.FakeClientCell(),
-		cell.Provide(func() *node.LocalNodeStore { return node.NewTestLocalNodeStore(node.LocalNode{}) }),
-		cell.Invoke(
-			func(
-				jg_ job.Group,
-				localNode_ k8s.LocalCiliumNodeResource,
-				localNodeStore_ *node.LocalNodeStore,
-				clientset_ *k8sClient.FakeClientset,
-			) {
-				jg = jg_
-				localNode = localNode_
-				localNodeStore = localNodeStore_
-				clientset = clientset_
-			},
-		),
-	)
-
-	tlog := hivetest.Logger(t, hivetest.LogLevel(slog.LevelError))
-	require.NoError(t, h.Start(tlog, t.Context()))
-	t.Cleanup(func() { h.Stop(tlog, context.Background()) })
-
-	_, err := clientset.CiliumV2().CiliumNodes().Create(t.Context(), initialNode, metav1.CreateOptions{})
-	require.NoError(t, err)
-
-	ipv4Allocator, ipv6Allocator, err := newENIMultiPoolAllocators(t.Context(), ENIMultiPoolAllocatorParams{
-		Logger:               hivetest.Logger(t),
-		IPv4Enabled:          true,
-		IPv6Enabled:          false,
-		CiliumNodeUpdateRate: time.Nanosecond,
-		Node:                 localNode,
-		LocalNodeStore:       localNodeStore,
-		CNClient:             clientset.CiliumV2().CiliumNodes(),
-		JobGroup:             jg,
-		Conf:                 &option.DaemonConfig{},
-	})
-	require.NoError(t, err)
-	require.NotNil(t, ipv4Allocator)
-	require.NotNil(t, ipv6Allocator)
-
-	assert.EventuallyWithT(t, func(c *assert.CollectT) {
-		updated, err := clientset.CiliumV2().CiliumNodes().Get(t.Context(), initialNode.Name, metav1.GetOptions{})
-		assert.NoError(c, err)
-		alloc := updated.Spec.IPAM.Pools.Allocated
-		require.Len(c, alloc, 1)
-		assert.Equal(c, defaults.IPAMDefaultIPPool, alloc[0].Pool)
-		assert.ElementsMatch(c, []iputil.Prefix{
-			iputil.PrefixFrom(eniCIDR1),
-			iputil.PrefixFrom(eniCIDR2),
-		}, alloc[0].CIDRs)
-	}, 5*time.Second, 10*time.Millisecond)
-}
-
-func TestDeriveENIVpcCIDRs(t *testing.T) {
-	t.Run("returns primary VPC CIDR from first ENI", func(t *testing.T) {
-		node := &ciliumv2.CiliumNode{}
-		node.Status.ENI.ENIs = map[string]awsTypes.ENI{
-			"eni-1": {
-				VPC: awsTypes.AwsVPC{
-					PrimaryCIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.0.0.0/16")),
-				},
-			},
-		}
-		primaryCIDR, secondaryCIDRs := deriveENIVpcCIDRs(node)
-		require.Equal(t, netip.MustParsePrefix("10.0.0.0/16"), primaryCIDR)
-		require.Empty(t, secondaryCIDRs)
-	})
-
-	t.Run("returns the secondary VPC CIDR associations", func(t *testing.T) {
-		node := &ciliumv2.CiliumNode{}
-		node.Status.ENI.ENIs = map[string]awsTypes.ENI{
-			"eni-1": {
-				VPC: awsTypes.AwsVPC{
-					PrimaryCIDR: iputil.PrefixFrom(netip.MustParsePrefix("10.1.128.0/19")),
-					CIDRs:       prefixes("100.64.0.0/16"),
-				},
-			},
-		}
-		primaryCIDR, secondaryCIDRs := deriveENIVpcCIDRs(node)
-		require.Equal(t, netip.MustParsePrefix("10.1.128.0/19"), primaryCIDR)
-		require.Equal(t, []netip.Prefix{netip.MustParsePrefix("100.64.0.0/16")}, secondaryCIDRs)
-	})
-
-	t.Run("returns zero when no ENIs", func(t *testing.T) {
-		node := &ciliumv2.CiliumNode{}
-		primaryCIDR, _ := deriveENIVpcCIDRs(node)
-		require.False(t, primaryCIDR.IsValid())
-	})
-
-	t.Run("returns zero when VPC CIDR is empty", func(t *testing.T) {
-		node := &ciliumv2.CiliumNode{}
-		node.Status.ENI.ENIs = map[string]awsTypes.ENI{
-			"eni-1": {
-				VPC: awsTypes.AwsVPC{},
-			},
-		}
-		primaryCIDR, _ := deriveENIVpcCIDRs(node)
-		require.False(t, primaryCIDR.IsValid())
-	})
-}
-
-func TestAutoDetectENINativeRoutingCIDR(t *testing.T) {
-	t.Run("auto-detects VPC CIDR when not configured", func(t *testing.T) {
-		logger := hivetest.Logger(t)
-		localNodeStore := node.NewTestLocalNodeStore(node.LocalNode{})
-
-		primaryCIDR := netip.MustParsePrefix("10.0.0.0/16")
-		conf := &option.DaemonConfig{}
-		require.NoError(t, autoDetectENINativeRoutingCIDR(logger, primaryCIDR, nil, localNodeStore, conf))
-
-		localNode, err := localNodeStore.Get(context.Background())
-		require.NoError(t, err)
-		require.True(t, localNode.Local.IPv4NativeRoutingCIDR.IsValid())
-		require.Equal(t, "10.0.0.0/16", localNode.Local.IPv4NativeRoutingCIDR.String())
-	})
-
-	t.Run("does not overwrite when already configured", func(t *testing.T) {
-		logger := hivetest.Logger(t)
-		localNodeStore := node.NewTestLocalNodeStore(node.LocalNode{})
-
-		primaryCIDR := netip.MustParsePrefix("10.0.0.0/16")
-		conf := &option.DaemonConfig{
-			IPv4NativeRoutingCIDR: netip.MustParsePrefix("10.0.0.0/8"),
-		}
-		require.NoError(t, autoDetectENINativeRoutingCIDR(logger, primaryCIDR, nil, localNodeStore, conf))
-
-		localNode, err := localNodeStore.Get(context.Background())
-		require.NoError(t, err)
-		// Should NOT have been written since the config already has a value.
-		require.False(t, localNode.Local.IPv4NativeRoutingCIDR.IsValid())
-	})
-
-	t.Run("accepts a native routing CIDR that is a subnet of the VPC CIDR", func(t *testing.T) {
-		// Regression test: a native routing CIDR that is a subset of the VPC
-		// CIDR (e.g. a single availability-zone subnet) is a valid, supported
-		// configuration. It must not be rejected.
-		logger := hivetest.Logger(t)
-		localNodeStore := node.NewTestLocalNodeStore(node.LocalNode{})
-
-		primaryCIDR := netip.MustParsePrefix("192.168.0.0/16")
-		conf := &option.DaemonConfig{
-			IPv4NativeRoutingCIDR: netip.MustParsePrefix("192.168.64.0/19"),
-		}
-		require.NoError(t, autoDetectENINativeRoutingCIDR(logger, primaryCIDR, nil, localNodeStore, conf))
-
-		localNode, err := localNodeStore.Get(context.Background())
-		require.NoError(t, err)
-		// Should NOT have been written since the config already has a value.
-		require.False(t, localNode.Local.IPv4NativeRoutingCIDR.IsValid())
-	})
-
-	t.Run("accepts a native routing CIDR matching a secondary VPC CIDR association", func(t *testing.T) {
-		// Regression test: pod subnets are commonly carved out of a secondary
-		// VPC CIDR association rather than the primary CIDR.
-		logger := hivetest.Logger(t)
-		localNodeStore := node.NewTestLocalNodeStore(node.LocalNode{})
-
-		primaryCIDR := netip.MustParsePrefix("10.1.128.0/19")
-		secondaryCIDRs := []netip.Prefix{netip.MustParsePrefix("100.64.0.0/16")}
-		conf := &option.DaemonConfig{
-			IPv4NativeRoutingCIDR: netip.MustParsePrefix("100.64.0.0/16"),
-		}
-		require.NoError(t, autoDetectENINativeRoutingCIDR(logger, primaryCIDR, secondaryCIDRs, localNodeStore, conf))
-
-		localNode, err := localNodeStore.Get(context.Background())
-		require.NoError(t, err)
-		// Should NOT have been written since the config already has a value.
-		require.False(t, localNode.Local.IPv4NativeRoutingCIDR.IsValid())
-	})
-
-	t.Run("rejects a native routing CIDR that overlaps no VPC CIDR", func(t *testing.T) {
-		logger := hivetest.Logger(t)
-		localNodeStore := node.NewTestLocalNodeStore(node.LocalNode{})
-
-		primaryCIDR := netip.MustParsePrefix("10.1.128.0/19")
-		secondaryCIDRs := []netip.Prefix{netip.MustParsePrefix("100.64.0.0/16")}
-		conf := &option.DaemonConfig{
-			IPv4NativeRoutingCIDR: netip.MustParsePrefix("192.168.0.0/16"),
-		}
-		require.Error(t, autoDetectENINativeRoutingCIDR(logger, primaryCIDR, secondaryCIDRs, localNodeStore, conf))
-	})
-}
